[fix - back] ranking 모듈 stats 모듈로 병합
