--- back/server/src/ranking/ranking.controller.spec.ts
@@ -1,20 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { RankingController } from './ranking.controller';
-import { RankingService } from './ranking.service';
-
-describe('RankingController', () => {
-  let controller: RankingController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [RankingController],
-      providers: [RankingService],
-    }).compile();
-
-    controller = module.get<RankingController>(RankingController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
